_cli, _errors: Move `VerificationError` to fix circular import
